Add mpich module.
We need this to have a different mpich from outside and inside the
config. eventually we don't need our own mpich from outsode and we could
use the one from the system to simplify.

But we still use our own mpich outside due to the bug we found in impi.